Issue #SB-15614 fix: de-norm job refactor
