+++ data-pipeline/de-normalization/src/main/java/org/ekstep/ep/samza/domain/Event.java
@@ -1,48 +1,24 @@
 package org.ekstep.ep.samza.domain;
 
 
-import com.google.gson.Gson;
+import org.ekstep.ep.samza.events.domain.Events;
 import org.ekstep.ep.samza.reader.NullableValue;
-import org.ekstep.ep.samza.reader.Telemetry;
 import org.ekstep.ep.samza.task.DeNormalizationConfig;
-import org.ekstep.ep.samza.util.Path;
 import org.joda.time.DateTime;
 import org.joda.time.format.DateTimeFormat;
 import org.joda.time.format.DateTimeFormatter;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.*;
 
-public class Event {
-    private final Telemetry telemetry;
-    private Path path;
+public class Event extends Events {
     private DateTimeFormatter dateFormat = DateTimeFormat.forPattern("yyyy-MM-dd").withZoneUTC();
     private DateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
 
     public Event(Map<String, Object> map) {
-        this.telemetry = new Telemetry(map);
-        path = new Path();
-    }
-
-    public Map<String, Object> getMap() {
-        return telemetry.getMap();
-    }
-
-    public String getJson() {
-        Gson gson = new Gson();
-        String json = gson.toJson(getMap());
-        return json;
-    }
-
-    public Long ets() {
-        NullableValue<Object> ets = telemetry.read("ets");
-        if(ets.value().getClass().equals(Double.class)) {
-            return ((Double) ets.value()).longValue();
-        }
-        return ((Long) ets.value());
+        super(map);
     }
 
     public Long getEndTimestampOfDay(String date) {
@@ -55,83 +31,39 @@ public class Event {
         Long currentMillis = new Date().getTime();
         String currentDate = formatter.format(currentMillis);
         Long endTsOfCurrentDate = getEndTimestampOfDay(currentDate);
-        if(eventEts > endTsOfCurrentDate) telemetry.add("ets", currentMillis);
+        if (eventEts > endTsOfCurrentDate) telemetry.add("ets", currentMillis);
         return ets();
     }
 
     public Boolean isOlder(Integer periodInMonths) {
         Long eventEts = ets();
         Long periodInMillis = new DateTime().minusMonths(periodInMonths).getMillis();
-        if(eventEts < periodInMillis) return true;
+        if (eventEts < periodInMillis) return true;
         else return false;
     }
 
-    public String eid() {
-        NullableValue<String> eid = telemetry.read("eid");
-        return eid.value();
-    }
-
-    public String did() {
-        NullableValue<String> did = telemetry.read("dimensions.did");
-        return did.isNull() ? telemetry.<String>read("context.did").value() : did.value();
-    }
-
-    public String channel() {
-        NullableValue<String> channel = telemetry.read("dimensions.channel");
-        return channel.isNull() ? telemetry.<String>read("context.channel").value() : channel.value();
-    }
-
-    public String objectID() {
-        if (objectFieldsPresent()) {
-            return telemetry.<String>read("object.id").value();
-        }
-        else return null;
-    }
 
     public String objectRollUpl1ID() {
-        if(objectFieldsPresent() && objectRollUpl1FieldsPresent()) {
-                return telemetry.<String>read("object.rollup.l1").value();
-        }
-        else return null;
+        if (objectFieldsPresent() && objectRollUpl1FieldsPresent()) {
+            return telemetry.<String>read("object.rollup.l1").value();
+        } else return null;
     }
 
-    public String objectType() {
-        if (objectFieldsPresent()) {
-            return telemetry.<String>read("object.type").value();
-        }
-        else return null;
-    }
-
-    public boolean objectFieldsPresent() {
-        String objectId = telemetry.<String>read("object.id").value();
-        String objectType = telemetry.<String>read("object.type").value();
-        return null != objectId && null != objectType && !objectId.isEmpty() && !objectType.isEmpty();
-    }
 
     public boolean objectRollUpl1FieldsPresent() {
-        if(objectFieldsPresent()) {
+        if (objectFieldsPresent()) {
             String objectrollUpl1 = telemetry.<String>read("object.rollup.l1").value();
             return null != objectrollUpl1 && !objectrollUpl1.isEmpty();
-        }
-        else return false;
+        } else return false;
     }
 
     public boolean checkObjectIdNotEqualsRollUpl1Id() {
         if (objectRollUpl1FieldsPresent() && !(objectID().equals(objectRollUpl1ID())))
             return true;
-         else
+        else
             return false;
     }
 
-    public String actorId() {
-        NullableValue<String> actorid = telemetry.read("uid");
-        return actorid.isNull() ? telemetry.<String>read("actor.id").value() : actorid.value();
-    }
-
-    public String actorType() {
-        NullableValue<String> actortype = telemetry.read("actor.type");
-        return actortype.value();
-    }
 
     public List<String> dialCode() {
         NullableValue<Object> dialcode = telemetry.read("edata.filters.dialcodes");
@@ -139,20 +71,17 @@ public class Event {
         if (dialcode != null && dialcode.value() != null) {
             if (dialcode.value().getClass().equals(String.class)) {
                 ids.add(dialcode.value().toString());
-            }
-            else {
+            } else {
                 ids = ((List) dialcode.value());
             }
-        }
-        else {
+        } else {
             NullableValue<Object> dialCode = telemetry.read("edata.filters.dialCodes");
             if (dialCode != null && dialCode.value() != null) {
                 telemetry.add("edata.filters.dialcodes", dialCode.value());
                 telemetry.add("edata.filters.dialCodes", null);
                 if (dialCode.value().getClass().equals(String.class)) {
                     ids.add(dialCode.value().toString());
-                }
-                else {
+                } else {
                     ids = ((List) dialCode.value());
                 }
             }
@@ -165,10 +94,9 @@ public class Event {
             case "device":
                 return did();
             case "user":
-                if("system".equalsIgnoreCase(actorType())) {
+                if ("system".equalsIgnoreCase(actorType())) {
                     return "";
-                }
-                else {
+                } else {
                     return actorId();
                 }
             case "content":
@@ -205,19 +133,18 @@ public class Event {
     public Map addISOStateCodeToDeviceData(Map deviceData) {
         // add new statecode field
         String statecode = deviceData.get("statecode").toString();
-        if(statecode != null && !statecode.isEmpty()) {
+        if (statecode != null && !statecode.isEmpty()) {
             String iso3166statecode = "IN-" + statecode;
             deviceData.put("iso3166statecode", iso3166statecode);
             return deviceData;
-        }
-        else return deviceData;
+        } else return deviceData;
     }
 
     public void addDeviceData(Map newData) {
         NullableValue<Map<String, Object>> previousData = telemetry.read(path.deviceData());
         Map<String, Object> deviceData = previousData.isNull() ? new HashMap<>() : previousData.value();
         // add new statecode field
-        if(!deviceData.isEmpty()) {
+        if (!deviceData.isEmpty()) {
             deviceData = addISOStateCodeToDeviceData(deviceData);
         }
         deviceData.putAll(newData);
@@ -230,8 +157,8 @@ public class Event {
         Map<String, Object> userdata = previousData.isNull() ? new HashMap<>() : previousData.value();
         userdata.putAll(newData);
         telemetry.add(path.userData(), userdata);
-        if(newData.size()>2)
-        setFlag(DeNormalizationConfig.getUserLocationJobFlag(), true);
+        if (newData.size() > 2)
+            setFlag(DeNormalizationConfig.getUserLocationJobFlag(), true);
         else
             setFlag(DeNormalizationConfig.getUserLocationJobFlag(), false);
     }
@@ -245,11 +172,11 @@ public class Event {
     }
 
     public void addCollectionData(Map newData) {
-            NullableValue<Map<String, Object>> previousData = telemetry.read(path.collectionData());
-            Map<String, Object> collectionData = previousData.isNull() ? new HashMap<>() : previousData.value();
-            collectionData.putAll(newData);
-            telemetry.add(path.collectionData(), collectionData);
-            setFlag(DeNormalizationConfig.getCollectionLocationJobFlag(), true);
+        NullableValue<Map<String, Object>> previousData = telemetry.read(path.collectionData());
+        Map<String, Object> collectionData = previousData.isNull() ? new HashMap<>() : previousData.value();
+        collectionData.putAll(newData);
+        telemetry.add(path.collectionData(), collectionData);
+        setFlag(DeNormalizationConfig.getCollectionLocationJobFlag(), true);
     }
 
     public void addDialCodeData(Map newData) {
@@ -258,6 +185,7 @@ public class Event {
     }
 
     public void setFlag(String key, Object value) {
+        System.out.println("rs" + path.flags());
         NullableValue<Map<String, Object>> telemetryFlag = telemetry.read(path.flags());
         Map<String, Object> flags = telemetryFlag.isNull() ? new HashMap<>() : telemetryFlag.value();
         if (!key.isEmpty()) flags.put(key, value);
@@ -273,13 +201,6 @@ public class Event {
     public String getUpgradedVersion() {
         BigDecimal updatedVer = BigDecimal.valueOf(Double.parseDouble(telemetry.read(path.ver()).value().toString())).add(BigDecimal.valueOf(0.1));
         return updatedVer.toString();
-    }
-
-    @Override
-    public String toString() {
-        return "Event{" +
-                "telemetry=" + telemetry +
-                '}';
     }
 
     public void markSkipped() {
@@ -302,9 +223,6 @@ public class Event {
         telemetry.add("metadata.src", config.jobName());
     }
 
-    public String edataType() {
-            return telemetry.<String>read("edata.type").value();
-    }
 
 }
